services/vehicle-service/requirements.txt
fastapi
uvicorn
redis
kafka-python
